+++ examples/sandbox/sandbox.py
@@ -36,16 +36,18 @@ netParams.sizeZ = 50
 
 
 # Population parameters
-netParams.popParams['PYR1'] = {'cellModel': 'HH', 'cellType': 'PYR', 'numCells': 10}#, 'xRange': [20,80], 'ynormRange': [0.3,0.7]}) # add dict with params for this pop 
+netParams.popParams['PYR1'] = {'cellModel': 'HH', 'cellType': 'PYR', 'numCells': 100}#, 'xRange': [20,80], 'ynormRange': [0.3,0.7]}) # add dict with params for this pop 
 #netParams.addPopParams('PYR2', {'cellModel': 'HH', 'cellType': 'PYR', 'numCells': 1}) # add dict with params for this pop 
-netParams.popParams['background'] = {'cellModel': 'NetStim', 'numCells': 500, 'rate': 20, 'noise': 0.5, 'start': 1, 'seed': 2}  # background inputs
+netParams.popParams['bkg1'] = {'cellModel': 'NetStim', 'numCells': 100, 'rate': 50, 'noise': 0.8, 'start': 1, 'seed': 2}  # background inputs
+netParams.popParams['bkg2'] = {'cellModel': 'IntFire2', 'numCells': 1, 'ib': 0}  # background inputs
+
 
 # Synaptic mechanism parameters
 netParams.synMechParams['AMPA'] = {'mod': 'Exp2Syn', 'tau1': 0.1, 'tau2': 1.0, 'e': 0}
 #netParams.addSynMechParams('esyn', {'mod': 'ElectSyn', 'g': 0.000049999999999999996})
 
-#netParams.stimSourceParams['background'] = {'type': 'NetStim', 'interval': 5, 'number': 1000, 'start': 600, 'noise': 0.1}
-#netParams.stimTargetParams['bkg->PYR1'] = {'source': 'background', 'conds': {'popLabel': 'PYR1'}, 'sec':'soma', 'loc': 0.5, 'weight': 0.5, 'delay': 1}
+netParams.stimSourceParams['background'] = {'type': 'NetStim', 'interval': 5, 'number': 10000, 'start': 100, 'noise': 0.5}
+netParams.stimTargetParams['bkg->PYR1'] = {'source': 'background', 'conds': {'popLabel': 'PYR1'}, 'sec':'soma', 'loc': 0.5, 'weight': 0.5, 'delay': 1}
 
 
 # Cell parameters
@@ -57,11 +59,21 @@ cellParams.conds = {'cellType': 'PYR'}
 netParams.addCellParams('PYR', cellParams)
 
 # Connections
-netParams.connParams['bg->PYR1'] = {
-    'preConds': {'popLabel': 'background'}, 'postConds': {'popLabel': 'PYR1'}, # background -> PYR
-    'weight': 0.1,                    # fixed weight of 0.08
-    'synMech': 'AMPA',                     # target NMDA synapse
-    'delay': 'uniform(1,5)'}           # uniformly distributed delays between 1-5ms
+# netParams.connParams['bkg1->PYR1'] = {
+#     'preConds': {'popLabel': 'bkg1'}, 'postConds': {'popLabel': 'PYR1'}, # background -> PYR
+#     'convergence': 10,
+#     'weight': 0.1,                    # fixed weight of 0.08
+#     'synMech': 'AMPA',                     # target NMDA synapse
+#     'delay': 'uniform(1,5)'}           # uniformly distributed delays between 1-5ms
+
+
+# netParams.connParams['PYR1->bkg2'] = {
+#     'preConds': {'popLabel': 'PYR1'}, 'postConds': {'popLabel': 'bkg2'}, # background -> PYR
+#     'probability': 0.2,
+#     'weight': 0.1,                    # fixed weight of 0.08
+#     'synMech': 'AMPA',                     # target NMDA synapse
+#     'delay': 'uniform(1,5)'}           # uniformly distributed delays between 1-5ms
+
 
 # netParams.addConnParams('PYR1->PYR2',
 #     {'preConds': {'popLabel': 'PYR1'}, 'postConds': {'popLabel': 'PYR2'}, # PYR1 -> PYR2 (gap junction)
@@ -87,7 +99,7 @@ simConfig.createPyStruct = 1  # create Python structure (simulator-independent) 
 simConfig.verbose = 0 #False  # show detailed messages 
 
 # Recording 
-simConfig.recordCells = []# [1,2]  # which cells to record from
+simConfig.recordCells = [] # [1,2]  # which cells to record from
 simConfig.recordTraces = {'Vsoma':{'sec':'soma','loc':0.5,'var':'v'}}
 
 simConfig.recordStim = False  # record spikes of cell stims
@@ -99,9 +111,9 @@ simConfig.saveFileStep = 1000 # step size in ms to save data to disk
 simConfig.savePickle = 0 # Whether or not to write spikes etc. to a .mat file
 
 # # Analysis and plotting 
-# simConfig.addAnalysis('plotRaster', {'spikeHist': 'subplot'})
+simConfig.addAnalysis('plotRaster', {})#, {'spikeHist': 'subplot'})
 # simConfig.addAnalysis('plotTraces', {'include': [0,1]})
-simConfig.addAnalysis('plot2Dnet', {'view': 'xz'})
+#simConfig.addAnalysis('plot2Dnet', {'view': 'xz'})