+++ CHANGES.md
@@ -6,7 +6,11 @@
 
 - analysis.plot2Dnet() 'view' param to select frontal ('xy') or top-down ('xz') view 
 
+- Now using noiseFromRandom123() for NetStims() (issue #89)
+
 - Fixed bug in readCmdLineArgs() 
+
+
 
 # Version 0.6.4